pyproject.toml
[project]
name = "rt-voice-spark-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.1",
    "gtts>=2.5.4",
    "huggingface-hub>=1.3.1",
    "llama-cpp-python>=0.3.16",
    "openai-whisper>=20250625",
    "piper-tts>=1.3.0",
]
